% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_functions.R
\name{countfreq}
\alias{countfreq}
\title{Counts observations by group including frequencies.}
\usage{
countfreq(x, ..., sort = FALSE, freq_digits = 1)
}
\arguments{
\item{x}{A tibble.}

\item{...}{Variables to group by (uses dplyr::count).}

\item{sort}{If TRUE will sort output in descending order of n.}

\item{freq_digits}{Number of digits to round percentages.}
}
\value{
A tibble with counts and frequencies.
}
\description{
This is based on dplyr::count, adding an extra column
with frequencies (rounded) within the lowest grouping level.
}
