% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_functions.R
\name{summarise_date}
\alias{summarise_date}
\title{Summarises dates.}
\usage{
summarise_date(x, ...)
}
\arguments{
\item{x}{A tibble.}

\item{...}{Comma separated list of unquoted expressions. You can treat variable names like they are positions. Use positive values to select variables; use negative values to drop variables. (This uses dplyr::select.)}
}
\value{
A tibble with the summary.
}
\description{
For grouped tibbles, summary statistics are done
for each group seperately. By default, analyses all numeric variables. You
can select specific variables using all syntax avaiable when using dplyr::select.
}
\details{
This selects all variables that are lubridate::is.instant().
}
