% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_functions.R
\name{summarise_na}
\alias{summarise_na}
\title{Counts missing observations (NA's)}
\usage{
summarise_na(x, ..., sort = TRUE, freq_digits = 1,
  remove_nonmissing = TRUE)
}
\arguments{
\item{x}{A tibble.}

\item{...}{Comma separated list of unquoted expressions. You can treat variable names like they are positions. Use positive values to select variables; use negative values to drop variables. (This uses dplyr::select.)}

\item{sort}{If TRUE will sort output in descending order of n.}

\item{freq_digits}{Number of digits to round percentages.}

\item{remove_nonmissing}{If TRUE remove variables without any missing values.}
}
\value{
A tibble with counts and frequencies.
}
\description{
This includes frequencies. By default, analyses all variables.
You can select specific variables using all syntax avaiable when using dplyr::select.
}
